requirements.txt
MetaTrader5>=5.0.45
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
packaging>=23.0
